pyproject.toml
[project]
name = "thaitravel"
version = "0.1.0"
description = ""
authors = [
    {name = "krahyor",email = "gimgim611@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "fastapi-cli (>=0.0.7,<0.0.8)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.24.0,<1.0.0)",
    "httpx (>=0.25.0,<1.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"